requirements.txt
requests==2.5.1
six==1.9.0
websocket-client==0.23.0
